refactor: HTTP 메서드 함수 위치 이동
